refactoring; fix-moved "propgagateSpawnedNodeChanges" from dashboard to spawner

+++ Tools/randomizer.go
@@ -31,9 +31,9 @@ func NewRandomizer(seed int64) *Randomizer {
 
 // Generates a random string of a given length and alphabet based on the randomizer seed.
 // Two calls to the same randomizer with the same parameters will usually not generate the same result.
-func (rand *Randomizer) GenerateRandomString(length int, validChars string) string {
+func (rand *Randomizer) GenerateRandomString(length uint32, validChars string) string {
 	str := ""
-	for i := 0; i < length; i++ {
+	for i := uint32(0); i < length; i++ {
 		rand.mutex.Lock()
 		str += string(validChars[rand.source.Int63()%int64(len(validChars))])
 		rand.mutex.Unlock()
@@ -43,18 +43,18 @@ func (rand *Randomizer) GenerateRandomString(length int, validChars string) stri
 
 // Generates a random number between min and max based on the randomizer seed.
 // Two calls to the same randomizer with the same parameters will usually not generate the same result.
-func (rand *Randomizer) GenerateRandomNumber(min, max int) int {
+func (rand *Randomizer) GenerateRandomNumber(min, max int64) int64 {
 	rand.mutex.Lock()
-	number := int(rand.source.Int63()%int64(max-min+1)) + min
+	number := rand.source.Int63()%int64(max-min+1) + min
 	rand.mutex.Unlock()
 	return number
 }
 
 // Generates a random string of a given length and alphabet based on a combination of the randomizer seed and the seed provided.
 // Two calls to the same randomizer with the same parameters will always generate the same result.
-func (randomizer *Randomizer) GenerateRandomStringWithSeed(length int, validChars string, seed int) string {
+func (randomizer *Randomizer) GenerateRandomStringWithSeed(length uint32, validChars string, seed int64) string {
 	str := ""
-	for i := 0; i < length; i++ {
+	for i := uint32(0); i < length; i++ {
 		str += string(validChars[(rand.NewSource(randomizer.seed+int64(seed)).Int63())%int64(len(validChars))])
 	}
 	return str
@@ -63,20 +63,20 @@ func (randomizer *Randomizer) GenerateRandomStringWithSeed(length int, validChar
 
 // Uses a combination of the randomizer seed and the seed provided.
 // Two calls to the same randomizer with the same parameters will always generate the same result.
-func (randomizer *Randomizer) GenerateRandomNumberWithSeed(min, max int, seed int) int {
-	return int(rand.NewSource(randomizer.seed+int64(seed)).Int63()%int64(max-min+1)) + min
+func (randomizer *Randomizer) GenerateRandomNumberWithSeed(min, max int64, seed int64) int64 {
+	return rand.NewSource(randomizer.seed+int64(seed)).Int63()%int64(max-min+1) + min
 }
 
 // RandomString generates a random string of a given length and alphabet.
 // Seed is the current system time.
 // Two calls to this function will usually not generate the same result.
-func RandomString(length int, alphabet string) string {
+func RandomString(length uint32, alphabet string) string {
 	return NewRandomizer(GetSystemTime()).GenerateRandomString(length, alphabet)
 }
 
 // RandomNumber generates a random number between min and max.
 // Seed is the current system time.
 // Two calls to this function will usually not generate the same result.
-func RandomNumber(min, max int) int {
+func RandomNumber(min, max int64) int64 {
 	return NewRandomizer(GetSystemTime()).GenerateRandomNumber(min, max)
 }
